Исправления после мержа

+++ Stars/app/src/main/java/com/se_au/stars/LeaderboardsProvider.java
@@ -19,7 +19,7 @@ public class LeaderboardsProvider {
         return mGlobalLeaderboardId;
     }
 
-    public void Submit(float result){
+    public void Submit(double result){
         if(!mGoogleApiClient.isConnected()){
             return;
         }
